Déplacer chargerDossiers de main.js vers action.js
